// apps/website/models/tagInventory.generated.ts
// Generated by yarn tags:inventory --write. Do not edit manually.
export const observedTags = [
  'academic',
  'competition',
  'extracurricular',
  'featured',
  'opensource',
  'org:appsbroker-consulting-limited',
  'org:fish-on-toast',
  'org:hackasoton',
  'org:hackscience',
  'org:hootsuite',
  'org:indemand',
  'org:isis-m3',
  'org:jsleague',
  'org:nicolae-grigorescu-national-college',
  'org:platform24',
  'org:r-ga',
  'org:stefanini',
  'org:university-of-southampton',
  'personal',
  'skills:alacritty',
  'skills:android',
  'skills:angular',
  'skills:angularjs',
  'skills:awk',
  'skills:aws',
  'skills:bash',
  'skills:c-c',
  'skills:casper-js',
  'skills:circleci',
  'skills:contentful',
  'skills:cordova',
  'skills:couchdb',
  'skills:crossref',
  'skills:css3',
  'skills:design-system',
  'skills:docker',
  'skills:electron',
  'skills:es6',
  'skills:eslint',
  'skills:express',
  'skills:figma',
  'skills:firebase',
  'skills:git',
  'skills:github-ci',
  'skills:gitlab-ci',
  'skills:go',
  'skills:google-apps',
  'skills:google-scholar',
  'skills:graphql',
  'skills:grpc',
  'skills:gunjs',
  'skills:heroku',
  'skills:html',
  'skills:html5',
  'skills:ios',
  'skills:javascript',
  'skills:javascript-es7',
  'skills:jenkins',
  'skills:jfrog',
  'skills:jotai',
  'skills:jsonschema',
  'skills:linux',
  'skills:lua',
  'skills:macos',
  'skills:make',
  'skills:mobx',
  'skills:mongodb',
  'skills:monorepo',
  'skills:moonrepo',
  'skills:mui',
  'skills:mysql',
  'skills:neo4j',
  'skills:netlify',
  'skills:nodejs',
  'skills:npm-registry',
  'skills:nvim',
  'skills:nx',
  'skills:onesignal',
  'skills:pdf',
  'skills:pdf2htmlex',
  'skills:phantom-js',
  'skills:php',
  'skills:postcss',
  'skills:prisma',
  'skills:python',
  'skills:raspberry-pi',
  'skills:react',
  'skills:react-hooks',
  'skills:react-native',
  'skills:recoil',
  'skills:redis',
  'skills:redux',
  'skills:resin-io',
  'skills:rest',
  'skills:rxjs',
  'skills:sass',
  'skills:sqlite',
  'skills:storybook',
  'skills:stylus',
  'skills:theming-system',
  'skills:tmux',
  'skills:trpc',
  'skills:turborepo',
  'skills:typescript',
  'skills:vagrant',
  'skills:vanilla-extract',
  'skills:vercel',
  'skills:vite',
  'skills:vue',
  'skills:webauthn',
  'skills:webgl',
  'skills:webpack',
  'skills:webrtc',
  'skills:websockets',
  'skills:wezterm',
  'skills:windows',
  'skills:yarn-2',
  'skills:yarn-3',
  'skills:yarn-4',
  'skills:yarn-plugins',
  'skills:yarn-workspaces',
  'skills:zsh',
  'year:2009',
  'year:2010',
  'year:2011',
  'year:2012',
  'year:2013',
  'year:2014',
  'year:2015',
  'year:2016',
  'year:2017',
  'year:2018',
  'year:2019',
  'year:2020',
  'year:2021',
  'year:2022',
  'year:2023',
  'year:2024',
  'year:2025',
  'year:2026',
] as const;

export type ObservedTagId = typeof observedTags[number];